+++ Infrastructure/Data/Migrations/EcommerceContextModelSnapshot.cs
@@ -35,7 +35,7 @@ namespace Infrastructure.Data.Migrations
                         .IsRequired()
                         .HasColumnType("TEXT");
 
-                    b.Property<decimal>("Price")
+                    b.Property<double>("Price")
                         .HasColumnType("decimal(20,2)");
 
                     b.Property<int>("ProductBrandId")
